chore: ajout d'un Makefile (#1638)
* chore: add a makefile

It simplifies the install procedure by using the classic make install
command
Since Make is well supported, shells provide auto complete for the
targets. Thus it simplifies starting the app with two targets:
start-front and start-back

* chore: rename make targets

Use app and backend instead of front and back. It's aligned with the
current naming.

* chore: use make in the installation documentation

* fix: ignore existing env files

cp -n does not override the existing env files. But it returns an error
which stops the install target. Adding a "-" at the start of the command
ignores those errors.

* chore: add make sub targets for app and backend

* chore: replace clean install by simple install for app

* chore(make): use cd instead of --prefix

* Update INSTALL.md

Co-authored-by: Jonathan Perret <jonathan.perret@beta.gouv.fr>

* chore(make): replace npm install by npm ci

* chore(make): add codegen and test targets

* Update Makefile

Co-authored-by: Jonathan Perret <jonathan.perret@beta.gouv.fr>

* Update Makefile

Co-authored-by: Jonathan Perret <jonathan.perret@beta.gouv.fr>

* Update Makefile

Co-authored-by: Jonathan Perret <jonathan.perret@beta.gouv.fr>

* Update Makefile

Co-authored-by: Jonathan Perret <jonathan.perret@beta.gouv.fr>

* feat(Makefile): pr review

* Update Makefile

---------

Co-authored-by: Jonathan Perret <jonathan.perret@beta.gouv.fr>